add tests for extra credit
